requirements.txt
absl-py==0.13.0
aioice==0.7.6
aiortc==1.3.2
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
astunparse==1.6.3
attrs==21.2.0
av==9.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
bleach==4.1.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
clang==5.0
click==7.1.2
colorama==0.4.4
colorlog==6.6.0
cryptography==37.0.2
cycler==0.11.0
Cython==0.29.30
debugpy==1.4.3
decorator==5.0.9
defusedxml==0.7.1
dnspython==2.2.1
entrypoints==0.3
Flask==2.0.1
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-crc32c==1.3.0
google-pasta==0.2.0
grpcio==1.40.0
h5py==3.1.0
htbuilder==0.6.0
idna==3.2
importlib-metadata==4.11.3
imutils==0.5.4
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.4
iteration-utilities==0.11.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mediapipe==0.8.10
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.11.0
notebook==6.4.3
numpy==1.19.5
oauthlib==3.1.1
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.7.0
pyee==9.0.4
Pygments==2.10.0
pylibsrtp==0.7.1
Pympler==1.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==304
pywinpty==2.0.5
pyzmq==22.2.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
semver==2.13.0
Send2Trash==1.8.0
six==1.15.0
smmap==4.0.0
st-annotated-text==3.0.0
streamlit==1.9.0
streamlit-lottie==0.0.3
streamlit-webrtc==0.37.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.0
typing-extensions==3.7.4.3
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.6
validators==0.18.2
vidgear==0.2.5
watchdog==2.0.3
wcwidth==0.2.3
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.8.0
pyttsx3
